refactor: remove optimize/ backward-compat shim package
All 11 test files now import directly from learning.optimize/.
The optimize/ directory was entirely composed of re-export shims.
