--- src/openjarvis/optimize/__init__.py
@@ -1,4 +1,0 @@
-"""Backward-compatibility shim -- optimize moved to learning.optimize."""
-
-from openjarvis.learning.optimize import *  # noqa: F401,F403
-from openjarvis.learning.optimize import __all__  # noqa: F401
