--- src/openjarvis/optimize/config.py
@@ -1,4 +1,0 @@
-"""Backward-compatibility shim -- optimize.config moved to learning.optimize.config."""
-
-from openjarvis.learning.optimize.config import *  # noqa: F401,F403
-from openjarvis.learning.optimize.config import __all__  # noqa: F401
