--- src/openjarvis/optimize/feedback/__init__.py
@@ -1,4 +1,0 @@
-"""Backward-compat shim: moved to learning.optimize."""
-
-from openjarvis.learning.optimize.feedback import *  # noqa: F401,F403
-from openjarvis.learning.optimize.feedback import __all__  # noqa: F401
