--- src/openjarvis/optimize/feedback/judge.py
@@ -1,7 +1,0 @@
-"""Backward-compat shim: moved to learning.optimize."""
-
-from openjarvis.learning.optimize.feedback.judge import *  # noqa: F401,F403
-from openjarvis.learning.optimize.feedback.judge import (
-    __all__,  # noqa: F401
-    _parse_score,  # noqa: F401
-)
