--- src/openjarvis/optimize/personal/__init__.py
@@ -1,4 +1,0 @@
-"""Backward-compat shim: moved to learning.optimize."""
-
-from openjarvis.learning.optimize.personal import *  # noqa: F401,F403
-from openjarvis.learning.optimize.personal import __all__  # noqa: F401
