--- src/openjarvis/optimize/personal/synthesizer.py
@@ -1,4 +1,0 @@
-"""Backward-compat shim: moved to learning.optimize."""
-
-from openjarvis.learning.optimize.personal.synthesizer import *  # noqa: F401,F403
-from openjarvis.learning.optimize.personal.synthesizer import __all__  # noqa: F401
