--- src/openjarvis/optimize/search_space.py
@@ -1,4 +1,0 @@
-"""Backward-compat shim: moved to learning.optimize."""
-
-from openjarvis.learning.optimize.search_space import *  # noqa: F401,F403
-from openjarvis.learning.optimize.search_space import __all__  # noqa: F401
