--- src/openjarvis/optimize/store.py
@@ -1,4 +1,0 @@
-"""Backward-compatibility shim -- optimize.store moved to learning.optimize.store."""
-
-from openjarvis.learning.optimize.store import *  # noqa: F401,F403
-from openjarvis.learning.optimize.store import __all__  # noqa: F401
