--- src/openjarvis/optimize/trial_runner.py
@@ -1,4 +1,0 @@
-"""Backward-compat shim: moved to learning.optimize."""
-
-from openjarvis.learning.optimize.trial_runner import *  # noqa: F401,F403
-from openjarvis.learning.optimize.trial_runner import __all__  # noqa: F401
