--- src/openjarvis/optimize/types.py
@@ -1,7 +1,0 @@
-"""Backward-compatibility shim -- optimize.types moved to learning.optimize.types."""
-
-from openjarvis.learning.optimize.types import *  # noqa: F401,F403
-from openjarvis.learning.optimize.types import (
-    _PARAM_TO_RECIPE,  # noqa: F401
-    __all__,  # noqa: F401
-)
