+++ tests/learning/test_feedback_collector.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 from unittest.mock import MagicMock
 
 from openjarvis.core.types import Trace
-from openjarvis.optimize.feedback.collector import FeedbackCollector
-from openjarvis.optimize.feedback.judge import TraceJudge
+from openjarvis.learning.optimize.feedback.collector import FeedbackCollector
+from openjarvis.learning.optimize.feedback.judge import TraceJudge
 
 
 def _make_trace(trace_id: str = "trace-001") -> Trace:
