+++ tests/learning/test_optimize_types.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from openjarvis.optimize.types import (
+from openjarvis.learning.optimize.types import (
     OptimizationRun,
     SampleScore,
     SearchDimension,
