+++ tests/learning/test_pareto.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 from typing import Any, Dict
 
 from openjarvis.evals.core.types import MetricStats, RunSummary
-from openjarvis.optimize.optimizer import compute_pareto_frontier
-from openjarvis.optimize.types import (
+from openjarvis.learning.optimize.optimizer import compute_pareto_frontier
+from openjarvis.learning.optimize.types import (
     ObjectiveSpec,
     TrialConfig,
     TrialResult,
