+++ tests/learning/test_search_space.py
@@ -2,11 +2,11 @@
 
 from __future__ import annotations
 
-from openjarvis.optimize.search_space import (
+from openjarvis.learning.optimize.search_space import (
     DEFAULT_SEARCH_SPACE,
     build_search_space,
 )
-from openjarvis.optimize.types import SearchSpace
+from openjarvis.learning.optimize.types import SearchSpace
 
 # ---------------------------------------------------------------------------
 # build_search_space
