+++ tests/learning/test_trace_judge.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from unittest.mock import MagicMock
 
 from openjarvis.core.types import StepType, Trace, TraceStep
-from openjarvis.optimize.feedback.judge import TraceJudge, _parse_score
+from openjarvis.learning.optimize.feedback.judge import TraceJudge, _parse_score
 
 # ---------------------------------------------------------------------------
 # _parse_score unit tests
